moved test cases over to new test file under the tests directory
